Consolidate import load
