+++ app/controllers/imports_controller.rb
@@ -1,5 +1,5 @@
 class ImportsController < ApplicationController
-  before_action :set_import, only: %i[ edit update destroy ]
+  before_action :set_import, except: %i[ index new create ]
 
   def index
     @imports = Current.family.imports
@@ -16,14 +16,14 @@ class ImportsController < ApplicationController
   def update
     @import.update!(account_id: params[:import][:account_id])
 
-    redirect_to import_load_path(@import), notice: "Import updated"
+    redirect_to load_import_path(@import), notice: "Import updated"
   end
 
   def create
     account = Current.family.accounts.find(params[:import][:account_id])
     @import = Import.create!(account: account)
 
-    redirect_to import_load_path(@import), notice: "Import was successfully created."
+    redirect_to load_import_path(@import), notice: "Import was successfully created."
   end
 
   def destroy
@@ -31,9 +31,25 @@ class ImportsController < ApplicationController
     redirect_to imports_url, notice: "Import was successfully destroyed.", status: :see_other
   end
 
+  def load
+  end
+
+  def update_csv
+    if @import.update(import_params)
+      redirect_to import_configure_path(@import), notice: "Import uploaded"
+    else
+      flash.now[:error] = @import.errors.full_messages.to_sentence
+      render :load, status: :unprocessable_entity
+    end
+  end
+
   private
 
     def set_import
       @import = Current.family.imports.find(params[:id])
+    end
+
+    def import_params
+      params.require(:import).permit(:raw_csv)
     end
 end
