+++ app/views/imports/_nav.html.erb
@@ -2,7 +2,7 @@
   <ul class="flex items-center gap-2">
     <%= render partial: "imports/nav_item", locals: { name: "Select", path: new_import_path, index: 0, is_complete: false, is_current: true } %>
     <div class="h-px bg-alpha-black-200 w-12"></div>
-    <%= render partial: "imports/nav_item", locals: { name: "Import", path: import.persisted? ? import_load_path(import) : nil, index: 1, is_complete: false, is_current: false } %>
+    <%= render partial: "imports/nav_item", locals: { name: "Import", path: import.persisted? ? load_import_path(import) : nil, index: 1, is_complete: false, is_current: false } %>
     <div class="h-px bg-alpha-black-200 w-12"></div>
     <%= render partial: "imports/nav_item", locals: { name: "Setup", path: import.persisted? ? import_configure_path(import) : nil, index: 2, is_complete: false, is_current: false } %>
     <div class="h-px bg-alpha-black-200 w-12"></div>
