+++ app/views/imports/load.html.erb
@@ -7,12 +7,11 @@
 <div>
   <h1 class="sr-only">Load import</h1>
 
-  <%= form_with model: @import, url: import_load_path(@import), method: :patch do |form| %>
+  <%= form_with model: @import, url: load_import_path(@import), method: :post do |form| %>
     <div class="mb-4">
-      <%= form.text_area :raw_csv, label: "CSV", value: @import.raw_csv, class: "", required: true %>
+      <%= form.text_area :raw_csv, label: "CSV", value: @import.raw_csv, required: true %>
     </div>
 
     <%= form.submit t(".next"), class: "px-4 py-2 block w-full rounded-lg bg-gray-900 text-white text-sm font-medium" %>
-
   <% end %>
 </div>
