+++ config/locales/views/imports/en.yml
@@ -23,9 +23,8 @@ en:
       imports: Imports
       new: New import
       title: Imports
-    loads:
-      show:
-        next: Next
+    load:
+      next: Next
     new:
       description_text: Importing transactions can only be done for one account at
         a time. You will need to go through this process again for other accounts.
